prices/producers/python/requirements.txt
autopep8==1.5.6
certifi==2020.12.5
chardet==4.0.0
flake8==3.9.1
idna==2.10
importlib-metadata==4.0.1
kafka-python==2.0.2
mccabe==0.6.1
oauthlib==3.1.0
pycodestyle==2.7.0
pyflakes==2.3.1
python-dotenv==0.17.0
requests==2.25.1
requests-oauthlib==1.3.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
zipp==3.4.1
